requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
weaviate-client==4.5.0
openai==1.12.0
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.6.1
langchain==0.1.9
tiktoken==0.6.0
numpy==1.26.4 